+++ src/flask_exts/forms/fields/datetime.py
@@ -3,8 +3,6 @@ import datetime
 from wtforms import fields as wtforms_fields
 from ..widgets.datetime import DateTimePickerWidget
 from ..widgets.datetime import TimePickerWidget
-from ...babel import _gettext
-
 
 class DateTimeField(wtforms_fields.DateTimeField):
     """
@@ -97,6 +95,6 @@ class TimeField(wtforms_fields.Field):
                     except ValueError:
                         pass
 
-                raise ValueError(_gettext("Invalid time format"))
+                raise ValueError(self.gettext("Invalid time format"))
             else:
                 self.data = None