+++ src/flask_exts/forms/validators/field_list.py
@@ -1,6 +1,4 @@
 from wtforms.validators import StopValidation
-from ...babel import _gettext
-
 
 class FieldListInputRequired:
     """Validates that at least one item was provided for a FieldList"""
@@ -10,4 +8,4 @@ class FieldListInputRequired:
     def __call__(self, form, field):
         if len(field.entries) == 0:
             field.errors[:] = []
-            raise StopValidation(_gettext("This field requires at least one item."))
+            raise StopValidation(field.gettext("This field requires at least one item."))