+++ CHANGES.rst
@@ -1,3 +1,13 @@
+v0.1.6
+------
+
+Released 2024-09-14
+
+- merge translations of wtforms, flask_exts and app into flask_babel.get_translations()
+- update FileField and ImageField
+- add url for templating.static
+- use namespace ns.bootstrap to control bootstrap's css and js
+
 v0.1.5
 ------